add: /register

+++ passguard/src/main/java/com/example/passguard/requests/user/get/GetUserRequest.java
@@ -3,14 +3,14 @@ package com.example.passguard.requests.user.get;
 import com.example.passguard.models.BaseRequest;
 
 public class GetUserRequest extends BaseRequest {
-    private final int id;
+    private final long id;
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
 
-    public GetUserRequest(int id, String token) {
+    public GetUserRequest(long id, String token) {
         super(token);
         this.id = id;
     }
